Se han ajustado algunas restricciones
